+++ src/MAPeD/src/py_scripting/py_api.cs
@@ -229,7 +229,7 @@ namespace MAPeD
 			
 			if( data != null )
 			{
-				return ( int )data.m_scr_ind;
+				return data.m_scr_ind;
 			}
 			
 			return -1;
